use float instead of double

+++ src/mpu6050.cpp
@@ -23,7 +23,7 @@ MPU6050::~MPU6050()
  * @param preferred_wire - pointer to the preferred wire object
  * @param g - acceleration due to gravity
 */
-MPU6050::begin(byte address , TwoWire* preferred_wire  ,double g )
+MPU6050::begin(byte address , TwoWire* preferred_wire  ,float g )
 {
     I2CDevice::begin(address,preferred_wire);
     _accelerometer_sensitivity = ACCELERO_METER_SENSITIVITY_2;
@@ -132,7 +132,7 @@ void MPU6050::disableSleepMode()
  * @note During calibration the sensor should be kept still
 */
 void MPU6050::calibrateSensor(int n){
-    double ax, ay, az, gx, gy, gz;
+    float ax, ay, az, gx, gy, gz;
     for (int  sample_no = 0; sample_no < n; sample_no++)
     {
         getSensorsReadings(ax, ay, az, gx, gy, gz,false);
@@ -161,13 +161,13 @@ void MPU6050::calibrateSensor(int n){
  * @param az - acceleration in z direction
  * @note The acceleration values are in m/s^2
 */
-void MPU6050::getAccelerometerReadings(double& ax, double& ay, double& az,bool calibrated)
+void MPU6050::getAccelerometerReadings(float& ax, float& ay, float& az,bool calibrated)
 {
     short int raw_a[3];
     readShortIntsFromReg(ACCEL_XOUT_H_REG, 3,  raw_a);
-    ax = (double) _acceleration_due_to_gravity * raw_a[0] * _accelerometer_sensitivity;
-    ay = (double) _acceleration_due_to_gravity * raw_a[1] * _accelerometer_sensitivity;
-    az = (double) _acceleration_due_to_gravity * raw_a[2] * _accelerometer_sensitivity;
+    ax = (float) _acceleration_due_to_gravity * raw_a[0] * _accelerometer_sensitivity;
+    ay = (float) _acceleration_due_to_gravity * raw_a[1] * _accelerometer_sensitivity;
+    az = (float) _acceleration_due_to_gravity * raw_a[2] * _accelerometer_sensitivity;
     ax -= calibrated ? _ax_offset : 0;
     ay -= calibrated ? _ay_offset : 0;
     az -= calibrated ? _az_offset : 0;
@@ -180,13 +180,13 @@ void MPU6050::getAccelerometerReadings(double& ax, double& ay, double& az,bool c
  * @param gz - angular velocity in z direction
  * @note The angular velocity values are in rad/s
 */
-void MPU6050::getGyroscopeReadings(double& gx, double& gy, double& gz,bool calibrated)
+void MPU6050::getGyroscopeReadings(float& gx, float& gy, float& gz,bool calibrated)
 {
     short int raw_g[3];
     readShortIntsFromReg(GYRO_XOUT_H_REG, 3,  raw_g);
-    gx = (double) raw_g[0] * _gyroscope_sensitivity;
-    gy = (double) raw_g[1] * _gyroscope_sensitivity;
-    gz = (double) raw_g[2] * _gyroscope_sensitivity;
+    gx = (float) raw_g[0] * _gyroscope_sensitivity;
+    gy = (float) raw_g[1] * _gyroscope_sensitivity;
+    gz = (float) raw_g[2] * _gyroscope_sensitivity;
     gx -= calibrated ? _gx_offset : 0;
     gy -= calibrated ? _gy_offset : 0;
     gz -= calibrated ? _gz_offset : 0;
@@ -202,7 +202,7 @@ void MPU6050::getGyroscopeReadings(double& gx, double& gy, double& gz,bool calib
  * @param gz - angular velocity in z direction
  * @note The acceleration values are in m/s^2 and the angular velocity values are in rad/s
 */
-void MPU6050::getSensorsReadings(double& ax, double& ay, double& az, double& gx, double& gy, double& gz,bool calibrated)
+void MPU6050::getSensorsReadings(float& ax, float& ay, float& az, float& gx, float& gy, float& gz,bool calibrated)
 {
     getAccelerometerReadings(ax, ay, az, calibrated);
     getGyroscopeReadings(gx, gy, gz, calibrated);
